+++ Algorithms.hpp
@@ -5,7 +5,7 @@ namespace ariel
     {
     public:
         static int isConnected(Graph g);
-        static string shortestPath(Graph g, int start, int end);
+        static string shortestPath(Graph g, size_t start, size_t end);
         static int isContainsCycle(Graph g);
         static string isBipartite(Graph g);
     };
